WIP refactor types
