+++ apps/app/src/interfaces/apiv3/index.ts
@@ -1,0 +1,1 @@
+export * from './page';
