+++ apps/app/src/interfaces/page-operation.ts
@@ -1,7 +1,3 @@
-import type {
-  IGrantedGroup, IPageHasId, IRevisionHasId, ITag, PageGrant,
-} from '@growi/core';
-
 export const PageActionType = {
   Create: 'Create',
   Update: 'Update',
@@ -30,33 +26,3 @@ export type IPageOperationProcessData = {
 export type IPageOperationProcessInfo = {
   [pageId: string]: IPageOperationProcessData,
 }
-
-export type OptionsToSave = {
-  isSlackEnabled: boolean;
-  slackChannels: string;
-  grant: PageGrant;
-  // userRelatedGrantUserGroupIds?: IGrantedGroup[];
-  // isSyncRevisionToHackmd?: boolean;
-};
-
-export type IApiv3PageCreateParams = {
-  path?: string,
-  parentPath?: string,
-  optionalParentPath?: string,
-
-  body?: string,
-  pageTags?: string[],
-
-  grant?: PageGrant,
-  grantUserGroupIds?: IGrantedGroup[],
-  overwriteScopesOfDescendants?: boolean,
-
-  isSlackEnabled?: boolean,
-  slackChannels?: string,
-};
-
-export type IApiv3PageCreateResponse = {
-  page: IPageHasId,
-  tags: ITag[],
-  revision: IRevisionHasId,
-};
